Context: detect symbol names in query (#62976)
This PR updates Cody context search to pull symbol names out of the query and
perform an additional symbol search. The symbol search results are deduplicated
with the code search results, so we always prefer the symbol definition over
another file match.

Benefits:
* When a query references a symbol, we always include its definition in context
(unless it is quite ambiguous). This is a basic thing users would expect, and
improves our "Find symbol" evals.
* We're more likely to return high quality chunks, since symbol results are
always anchored on the line with the definition.

Addresses SPLF-88