chore(appliance): fix race condition in tests
We observed a rare test flake in
TestNonNamespacedResourcesRemainWhenDisabled in which non-namespaced
resources were apparently not deleted when assertions were made against
golden fixtures. My theory for why this was happening, is that there was
a race between the test goroutine and the reconciler running in the
background.

The reconciler actually appears to emit 2 events per configmap creation.
The second is probably a result of updating the currentVersion
annotation at the end of the reconcile loop, which itself triggers
another reconcile loop.

Our test code was waiting for the initial resources to be created by
waiting for at least one reconcile event to appear. It was then waiting
for only one more event. When we got unlucky, this event would be the
2nd event from the 1st configmap upload, rather than the _next_ event,
from the 2nd configmap upload. If the test thread scraped all resources
from kube-apiserver before the reconciler had deleted them (in response
to the 2nd configmap upload in this particular test), the golden
assertions would fail.

This could all be totally wrong. We only received one report of such a
flake, and I was unable to reproduce it in hundreds of local runs, so
sadly I haven't empirically validated this.