gitserver: All syncers use the Git syncer
The Git syncer is the one true implementation we want to trust. It knows how to grab a single pack file, and it will in the future be the code path that can trigger ref update hooks etc.
Thus, all syncers should go through that layer.
Instead of having separate syncers, I decided to make just one second type of syncer: A Generating Syncer. It runs some method in a temp git dir to create the contents, and that's it. Then, we will instruct the git syncer to fetch the latest from that repo (via a file:// URL).

Test plan:

Manually tested that p4 fusion repos still clone alright. Also CI passes.