repoupdater(scheduler): Prevent race condition of schedule preloading
My local instance has few repos enough that this doesn't happen, but on larger instances this preloading fights with the new preloading.
They are both best effort, and are meant to achieve the same thing.
Thus, this one is not required anymore, and we can delete it, after we added another one in https://github.com/sourcegraph/sourcegraph/pull/62891.

Test plan:

Verified with sleeps and logs locally that repos are correctly upserted in the schedule now.