Initialise authProviders array in Go code (#63099)
When a Sourcegraph license is expired or revoked and the instance
reverts to free license, then `authz` (for permissions syncing) stops
working as well. This means that there are no AuthProviders set up and
the front-end code has `null` for these values.

The issue arrises from unitialised arrays in Go being marshalled as
`null` instead of `[]`. This change initialises the array explicitly,
instead of relying on the empty value.