Load endpoint URL before access token
`requestGraphQLFromVSCode`` accepts the endpoint URL
and access token as parameters, loading each from
storage if they are not provided.
The order in which those two parameters are
resolved is important because the key to the
access token is the endpoint URL. The resolution
order was backward: it first loaded the storage for
the access token (`session`), using the stored
endpoint URL (ignoring the passed-in value) and
then resolved the endpoint URL.
This resulted in a mismatch between the endpoint URL
and the access token, requiring the user to manually
edit the settings.json file to specify those two
parameters.

There are still problems with the auth process,
but now it will actually use the values entered
in the UI.