scim: Fix random suffix added to every user on resync (#63122)
In the current implementation, we check if there's a user in the
database already, and if so we add a random suffix to increase the
chances of the insert/update to succeed.

However, we didn't check if the user that already exists is _the same
user_. So for a first sync, usernames would look nice and tidy, but then
on a second sync from the SCIM provider, every user would get a suffix.

This PR fixes that by adding a check for the user ID.

Test plan:

Added tests, verified SCIM sync doesn't modify usernames anymore
locally.

(cherry picked from commit 614375e53869e37cf42f0db3669043c2f66849a9)