Use `endpointSetting()` to get the endpoint URL
Instead of directly accessing `vscode.workspace.getConfiguration('sourcegraph').get<string>('url')`.

This will allow for replacing where the endpoint URL is stored, and also, it may be that using `getConfiguration('sourcegraph')` to get the section of the config file was not working correctly. I was seeing some strange behavior with `readConfiguration()`, which does this internally.