scim: Fix user updates when SCIM was previously enabled
When the user external account entry for SCIM has been deleted in the past, updating the user again fails as the deleted record is found and then the insertion fails.

This PR fixes that by adjusting the unique index to only consider non-soft-deleted external accounts.

Test plan:

Added a test suite for the DB method - it was untested before :(