fix(search): Properly handle URLs with file paths that contain route segments (#63123)
Fixes SRCH-484. See that issue for more information about the problem.

Instead of hardcoding the routes that can contain file paths, I'm using
the generated list of routes in `routes.ts` and extract any route that
ends with a `[...path]` segment (which, so far at least, is used as the
file path parameter).

This list is then used to hook into the routing process and encode the
file path when any of the routes match. As mentioned in the `reroute`
documentation, this does not affect the URL displayed in the browser.
Because the file path is now URL encoded we need to decode it before we
can process it further (I grepped for all occurrences of `params.path`).

## Test plan

Manual testing, integration test.