export the extension context from `activate`
It is a literal export of the context, and I'm thinking there has to be a better way to do it.
It's used only in `endpointSetting.js` currently; there may be a way to register it there, but for now, this works.

The context is used to access the globalState storage, in which the endpoint URL is (will be, in a subsequent commit) stored.

Exporting the context, rather than `extensionCoreAPI` or `localStorageService` because it's the simplest right now.