fix(appliance): use version for image tags (#63075)
Simultaneously fix a bug in the (still unreleased) appliance and change
approach to container image metadata. The appliance used to maintain a
map of versions to image-tags. The only version hardcoded for the
still-in-development appliance was 5.3.9104, since that was the latest
at the time development started. These versions were set to 5.3.2, which
is incorrect. These versions were obtained from looking at the main
branch of deploy-sourcegraph-helm, instead of the relevant release
branch.

The change in approach is to stop including the optional sha256 digest
of the image in the appliance-generated config. We considered several
approaches in

https://linear.app/sourcegraph/issue/REL-22/spike-sketch-out-how-appliance-developers-will-handle-upgrades,
please see that issue for more context.

It appears simpler to statelessly determine the image tag (as being
equal to the release version string) rather than add moving parts such
as calling the release registry, or building a 2-phase release system to
bake image checksums into the appliance once they are known post-build.
These options introduce complexity and greater possibility of errors
than just asking for a version.

Closes

https://linear.app/sourcegraph/issue/REL-135/appliance-knows-what-images-to-pull-for-a-given-release.