Svelte: allow popovers to have external targets (#63141)
This updates the `Popover` component to accept a passed-in target rather
than setting the target via an action. This allows us to pass in targets
not owned by the `Popover` component, which is important for things like
the file tree where the logical target is a wrapper around what the
consumer has access to.