webhooks: Update payload testdata and type for Bitbucket Server push events
The plugin did not actually support push events yet, so this was newly added.
Not sure where the existing testdata came from, maybe the bitbucket native webhooks that don't support global scope, but this event was never handled right.
To make sure we actually support the payload returned from the plugin, this PR adjusts the payload.

Test plan:

CI passes with the new testdata, and with the updated event type.