webhooks: Add support for secrets for Bitbucket Cloud
We didn't support secrets for bitbucket cloud so far, but the code host does support them. They use the same mechanism as Bitbucket and GitHub so that was easy to add.

Test plan:

Set up a webhook locally and tested that with a wrong secret it fails, and with the correct secret it passes. Also wrote a test.