fix/internal/memcmd: fix goroutine leak in linux observer (#63206)
There was a goroutine leak in the Linux observation logic. 


https://github.com/sourcegraph/sourcegraph/blob/add4baa455ef86d60b59f7c0182839a7bb2a9e18/internal/memcmd/observer_linux.go#L122-L141

In observe, it was possible for the goroutine that produces collection
events on a channel to block forever. If we exit early from the
function, it was still possible for the channel to be full (because the
consumer exited early), which causes the producer goroutine to block
forever since there is no room in the channel.

This PR fixes this issue by adding a `defer` statement that ensures that
collection channel is drained before we exit - thus fixing the leak.



## Test plan

Added goroutine leak dector to linux tests - now see them pass.

## Changelog 

A goroutine leak in the experimental linux memory observation logic has
been fixed.