feat/gitserver: don't log memoryObservation error if it occured b/c context cancellation (#63210)
This PR changes the behavior of the git cli commands to ignore errors
from the memory observer if they occurred due to cancellation. Such
errors are expected if the user cancels the request, and so there is no
reason to generate logspam.

## Test plan

Existing CI tests

This PR only reduces logspam, and as such doesn't need extensive new
unit tests.


## Changelog

- Reduce logspam in the gitserver command functionality by ignoring
memory observation errors if they occurred due to external context
cancellation.