chore: Add GitBlobResolver and GitTreeResolver for readability (#63223)
In our backend, we mostly follow the convention that type names
for the resolvers match up with the types in the GraphQL API,
with the extra `Resolver` suffix. However, this is broken by the
`GitTreeEntryResolver` where there were no types called
`GitBlobResolver` or `GitTreeResolver`, since all functionality
was implemented on top of `GitTreeEntryResolver`.

Since Go supports implicit method forwarding for embedded structs,
I think we can improve the code consistency & readability by adding
these two stub resolver types.

As struct embedding is not a super common pattern, I've explicitly
mentioned that in the doc comments for `GitBlobResolver` and
`GitTreeResolver`.