chore(ci): conditionally stamp genrules (#63204)
See https://github.com/bazelbuild/bazel/releases/tag/7.1.0 which
introduces a new value for the stamp attribute for `genrule`: `stamp =
-1` which makes the stamping conditional.

Previously, any genrule that had `stamp = 1` would be stamping all the
time due to how it was implemented prior 7.1. An as we stamp builds with
a computed version that is date dependent, we would never get cache hits
on those, leading to `//cmd/migrator:image_test` being executed all the
time. It wasn't that big of an issue as it's rather quick, but still.

Now the `genrule` supports that there's no reason to fix that one. 


## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

CI 

## Changelog

- Avoid invalidating cache in stamped genrules on dev builds.