feat(search/svelte): Support more file icons (#63181)
Closes SRCH-452

Until this commit both the React and Svelte apps have used the same code
to determine the correct file icon. That code however still held
separate logic for the two apps with the result that our file icon set
for the Svelte app was a lot smaller.
    
This commit separates the logic and now makes use of `unplugin-icons` in
the Svelte app.
    
I tried to use `unplugin-icons` in the React so that the code could
continued to be shared but I couldn't get it to work.

Note that I didn't use the exact same icons as in the React app. I
wanted to minimize the number of different icon sets we are using.

## Test plan

Builds without error, manual testing.