fix(search): Copy text from read-only code host configuration input
PR #49682 made some CSS changes to hide the caret as well as bracket
matching for this input. I assume that seeing those in a read-only input
was confusing.

However, hiding the textarea element also breaks copying the input's
value. Instead of hiding the textarea, we are now hiding the visiual
representation of the caret.

A noteable difference in behavior now is that it's possible to focus the
input via tab, but there is no visual indiciation that the editor has
focus (since we hide the caret, as before).