Svelte: use Roboto Mono as monospace font (#63247)
This adds Roboto Mono as the default monospace font for the Svelte
webapp. Previously, we were nominally using SF Mono, but that is not
installed by default and we were not fetching it, so the figma designs
looked different than what was being rendered in the web app. This uses
FontSource packages for the font for easy self-hosting (no external
requests).