feat(search/svelte): Add context specific suggestions to repo search input (#62880)
This PR adds a new suggestions source to the repository search input. It
currently offers the following suggestions:

- Repository group/org (on every repo page)
- Current file and/or directory (on blob and tree pages)
- Language of current file (on blob pages)

When on a commit page it adds the revision of the commit to the repo
filter.

A new context store is introduced to propagate those values to the
search input. I also had to update our query processing logic to
preserve pattern and filter token ranges (first commit) so that we can
also replace filters, not just add the.

NOTE: I'm not convinced the way the information is propagated via
context is the best way to do, but it's the simplest I could think of
now. I can imagine in the future that individual pages create/register
suggestions directly.



https://github.com/sourcegraph/sourcegraph/assets/179026/7c1b78e3-061b-4dc8-a1ee-fb3273e633ba



## Test plan

Manual testing.

---------

Co-authored-by: Camden Cheek <camden@ccheek.com>