fix(svelte): Fix icon color in search sidebar
When selecting a file filter in the search sidebar the file icon doesn't
change color and is often difficult to see against the selected
background color.

This commit fixes that an cleans up the icon coloring logic to be more
consistent: All icons now have the default icon color `--icon-color`,
which is set globally. This color can be overridden by setting
`--icon-color` "locally", i.e. closer to the icon.