fix/graphql/database: make count of failed user permissions syncs ignore soft deleted users (#63297)
This PR changes the query that powers the "failed user sync count" on
the permissions center to now properly ignore soft-deleted users. The
admin can't take any action to remedy this, so there is no point in
displaying it.


![Screenshot 2024-06-17 at
12.52.08 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/5VKJ5spRdhDRvKQ0TTIe/6ef49bc9-fc65-4ef9-a58e-f2a66ff87e31.png)


## Test plan

Unit tests

## Changelog 

- The failed user permission sync count on the permission dashboard now
properly ignores syncs from deleted users (which is the expected
behavior).