fix/graphql/database: fix repo permission sync count to ignore soft deleted repositories (#63298)
This PR changes the query that powers the "failed repository sync count"
on the permissions center to now properly ignore soft-deleted
repositories. The admin can't take any action to remedy this, so there
is no point in displaying it.

![Screenshot 2024-06-17 at
12.55.19 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/5VKJ5spRdhDRvKQ0TTIe/ca554b82-c663-421b-8208-5f8c52a1c6ab.png)



## Test plan

Unit tests

## Changelog 

- The failed repository permission sync count on the permission
dashboard now properly ignores syncs from deleted repository (which is
the expected behavior).