fix/frontend/graphql: log and add trace events for all non-not-found errors that occur when fetching permission syncs (#63302)
Closes
https://linear.app/sourcegraph/issue/SRC-421/customer-reported-an-issue-with-user-permissions-stats

This PR ensures that any non not-found errors that occur when fetching
permission sync jobs are logged and added as error events in the trace.
Before, these errors were silently swallowed which makes things
exceptionally hard to debug.

This PR only changes the monitoring around this query - **it doesn't
change the underlying behavior.**

**Note**: I spent a while reading up on how to communicate partial
errors in our graphql API, but it seems like there isn't a clear
consensus.
- https://github.com/graph-gophers/graphql-go/issues/414
- https://spec.graphql.org/October2021/#sel-HAHlBNJDPEBAAADFBU-_D

According to the above, it seems like ideally we should be returning
`null` in the graphQL for permission sync jobs that have an error.
However, when I tried to do this I begain seeing nil panics all over the
place. It's hard for me to know if the underlying library supports this
or if we are just holding it wrong. 😥

## Test plan

Unit tests

## Changelog

- Our graphqlAPI now logs and traces any non-not-found errors that occur
when fetching permission sync jobs (as opposed to being silently
swallowed).