fix(svelte): Fix reference panel loading/error state (#63316)
Contributes to srch-578

Currently we keep showing the loading indicator even if no references
have been found. This commit moves the info messages into the layout
because the layout has information we need to properly distinguish the
different states.

I think more refactoring is necessary to make this API better but that
will be done as part of adding search-based code intel support.


## Test plan

Manual testing.