feat/enterpriseportal: add Salesforce metadata at subscription level (#63317)
Today Salesforce "subscription ID" and "opportunity ID" are attached at
the license level. They should really be on the Enterprise subscription
level instead, which are parent entities of licenses. As we move forward
with CORE-149, I think this is a good chance to introduce this change in
our schema, and migrate the currently per-license information to the
subscription level when we move things into Enterprise Portal.

## Test plan

CI