wip(cody): Introduce `cody-gateway-config` tool for generating model configuration (#63324)
This PR adds a couple of foundational components of the server-side LLM
configuration project. ([Server-side Cody Model
Selection](https://linear.app/sourcegraph/project/server-side-cody-model-selection-cca47c48da6d))

1. A small command-line tool named `cody-gateway-config`, which will
write a JSON file to disk.
2. The `internal/modelconfig` package for defining LLM Model
Configuration data.

## What this enables

The `cody-gateway-config` contains logic for building a
`ModelConfiguration` object in-memory, and saving it to disk. The
resulting JSON file is intended to be persisted in the repository with
the name `cody-gateway-model-configuration.json`.

The purpose of this tool is to track the current list of LLM models that
Cody Gateway supports in-code, so it can be versioned as part of the
regular code-review and PR process. (Generally, this is just a fancier
way to generate the same data that is in the [`sourcegraph/cody` repo's
`dotcom.ts`
file](https://github.com/sourcegraph/cody/blob/4f70bff5fea368b5febe11972567794f553e42f5/lib/shared/src/models/dotcom.ts#L4).)

With the file being checked into the repo, we can then statically embed
it into any Golang binaries that reference the
`internal/modelconfig/embedded` package. This is to make it super-easy
to distribute the latest Cody Gateway configuration data, since we do
not need to copy the file and/or figure out how to load it from the
local disk. And instead, just have the read-only contents available
in-memory. (If size ever becomes a concern, we could embed a compressed
version, etc.)

## Why that is needed

After this PR gets merged, both Cody Gateway and the Sourcegraph
instance codebases will be able to take a dependency on this
`modelconfig/embedded` package and use this data as part of the larger
"server-side LLM model configuration" story.

- Cody Gateway will simply add a new HTTP endpoint like
https://cody-gateway.sourcegraph.com/llm-models.json and just serve the
raw data.
- Sourcegraph instances will have static knowledge about available LLM
models (based on when they were built/released), and be able to poll
Cody Gateway to pick up on any updates.
- Sourcegraph instances will then be able to customize/overwrite/extend
this configuration data in arbitrary ways. (e.g. to support BYOK or
BYOLLM.)
- Sourcegraph instances will also expose their own "view" of this data
to Cody clients. So that Cody clients can fetch a list of "known
models", to allow users to select between various options, etc.

There's a lot that I'm glossing over in that list. But the big picture
is that we will use these data types, and static `.json` file to power
how this information will flow between various Sourcegraph components.

## Test plan

Added unit tests for validating the generated configuration data is
valid.

## Changelog

NA, this is an initial step. A later PR will elaborate on the exciting
new feature this is enabling.