feat/enterpriseportal: CreateEnterpriseSubscriptionLicense RPC defininition (#63323)
Defines an RPC for creating licenses for a subscription in Enterprise
Portal (which will replace the existing
`generateProductLicenseForSubscription` GraphQL mutation). Because of
#63317 we will no longer accept Salesforce details as part of the input
for creating licenses. Most everything else stays the same.

https://google.aip.dev/133 says we must name the parent subscription ID
as the `parent` parameter. But, we already have a `subscription_id` in
`EnterpriseSubscriptionLicense`, which feels more descriptive, so I've
opted not to use `parent`... but I'm open to suggestions otherwise 🤷

Closes CORE-150, implementation tracked in CORE-157

## Test plan

CI