fix(svelte): Preserve selected lines in codehost URL
It was reported that the new web app doesn't preserve the selected lines
when navigating to the corresponding file in the code host. This commit
fixes that.

Additionally this commit adds file actions (most importantly the 'view
in codehost' one) to the 'file at revision' view.

And while working on that I also fixed the 'view at commit' button in
the history panel, which should not show 'close commit' when the inline
diff view is open.