fix(svelte): History suggestions should show immediately when clicking the history button (#63335)
Fixes srch-545

We are hiding suggestions until the user has interacted with the input
(e.g. by clicking or typing), but, as it is not, it does not include
clicking on the history button.

This commit changes that.


## Test plan

Manual testing. Clicking the history button without having interacted
with the search input in any other way will now show history suggestions
and focus in the input.