chore(svelte): Simplify development for dotcom (#63341)
Whether or not the app is in "dotcom mode" is determined by
`window.context.sourcegraphDotComMode`. In development we use a stub
`window.context` object defined in `app.html`.

This commit extends that object to set `sourcegraphDotComMode` depending
on an environment varible. The environment variable is set by using vite
modes (https://vitejs.dev/guide/env-and-mode.html).

The whole setup can be started with `pnpm dev:dotcom`.


## Test plan

Ran `pnpm dev:dotcom` and saw the dotcom specific header and footer.