Refactor the 'getModel' callbacks into their own file (#63359)
Minor refactoring to the `cmd/frontend/internal/completions` package.

When we call `newCompletionsHandler` one of the parameters is a function
named `getModel`. This is called to determine which LLM model should be
use to respond to the incoming completion request. And we have two
implementations of this function, one for code completions and another
for chats.

This PR just moves the logic for those two implementations into their
own file (`get_model.go`), along with a couple of functions for which
they were the only caller.

There were two minor functionality changes I made, which I'll call out
in comments on this PR.

## Why?

As we rework how LLM models and associated configuration flows
throughout the backend, updating these functions will be a bit easier if
they are pulled out like this rather than being defined inline like they
are today.

Also, pretty much any place in the codebase where we have hard-coded an
LLM model is "on notice" and should instead be driven entirely by some
sort of global configuration file. (Since this is one of the key
problems server-side LLM config is trying to solve.)

## Test plan

NA, no functional changes. Relying on CI/CD and linter.

## Changelog

NA